requirements.txt
django==2.2.13
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-environ==0.4.5

black==20.8b1
pylint==2.7.0
pytest==6.2.2